Appendix C1 Berechnung fertig und g mit eingebaut

+++ TRPO/policy.py
@@ -27,12 +27,12 @@ class NN(object):
 
 
     def q(self, s, a):
-        mu = self.model(torch.tensor(s)).detach().numpy() # Vektor für den action space
+        mu = self.model(torch.tensor(s, dtype = torch.float)).detach().numpy() # Vektor für den action space
         dev = np.exp(self.log_dev.detach().numpy()) # deviation
         return norm.pdf(a, mu, dev)
 
     def ableitbar_q(self, s, a): #liefert das gleiche zurück wie q nur für torch interpretierbar, sodass diese Funktion optmiert werden kann
-        mu = self.model(torch.tensor(s)).double()
+        mu = self.model(torch.tensor(s, dtype = torch.float)).double()
         dev = torch.exp(self.log_dev).double()
         covariance_matrix = torch.eye(dev.shape[0]).double()*dev
         # factor = 1/(torch.tensor(np.sqrt(2*np.pi))*dev)
